+++ MegalomaniaPlugin/Utilities/Utils.cs
@@ -1,5 +1,6 @@
 ﻿using BepInEx;
 using MegalomaniaPlugin.Skills;
+using MegalomaniaPlugin.Skills.MinigunAbility;
 using RoR2;
 using RoR2.Orbs;
 using RoR2.Skills;
@@ -78,6 +79,7 @@ namespace MegalomaniaPlugin.Utilities
             SkillLookup.Add("bomb", BombAbility.BombSkill);
             SkillLookup.Add("twinshot", TwinShotAbility.TwinShotSkill);
             SkillLookup.Add("shell", ShellAbility.ShellSkill);
+            SkillLookup.Add("minigun", FireMinigun.MinigunSkill);
         }
 
 #nullable enable
@@ -188,6 +190,14 @@ namespace MegalomaniaPlugin.Utilities
 
                 inventory.RemoveItem(toTransform);
                 inventory.GiveItem(toGive);
+
+                if (MegalomaniaPlugin.ConfigRegenerateScrap.Value)
+                {
+                    if (toTransform == DLC1Content.Items.RegeneratingScrap.itemIndex)
+                    {
+                        inventory.GiveItem(DLC1Content.Items.RegeneratingScrapConsumed);
+                    }
+                }
 
                 //balance transformation over time
                 if (!ignoreCap)